🗓️ Temporarily Removed Sunrise RVA Calendar due to Event Reduplication
Sunrise RVA consistently posted many and repeating events from other calendars and groups, which is good, but outside the scope of what the calendar's purpose of rva.rip ultimately is (and duplicated many of the entries present on rva.rip)

If you're with Sunrise RVA, please don't hesitate to reach out to me to work on a potential solution. Your movement is great and I'd love to continue having your public events.